+++ changelog.md
@@ -14,7 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Markdown plugins
 - Platforms content and component
 - Portfolio configuration, content and functionality
-
+- Experience configuration, content and functionality
+- 
 ## [0.0.0] - 2019-01-14
 
 ### Added